Refactor discord bot
